requirements.txt
uvicorn==0.34.0
python-dotenv==1.1.0
fastapi==0.115.12
httpx==0.28.1
sentence-transformers==2.2.2
pinecone==6.0.2 
python-docx==1.1.2
transformers==4.35.2
pymongo==4.12.0
passlib==1.7.4
motor==3.7.0
bcrypt==3.2.0
email-validator==2.2.0
google-generativeai==0.8.5